Move interop element measure policy to interop element layout
